Temporarily Undo Previous Version
We discovered that when you click on the checkbox of GemCore on the server mods tab via the In-game Mod Manager, the game will crash because of redundant meta function definitions in gemscripts/tableutils.lua that are already defined in DST's util.lua. Also, the meta functions in tableutils.lua were incorrectly re-implemented.